[FIX] udes_stock: Disable name translation
Disable name translation for picking types and routes,
this avoids complications when duplicating, where the
new duplicated name was not being saved and was being
treated as a translation.

User-story: 621
