+++ addons/udes_stock/models/__init__.py
@@ -13,6 +13,7 @@ from . import stock_inventory
 from . import stock_location
 from . import stock_location_category
 from . import stock_location_path
+from . import stock_location_route
 from . import stock_move
 from . import stock_move_line
 from . import stock_picking
@@ -23,3 +24,4 @@ from . import stock_production_lot
 from . import stock_quant
 from . import stock_quant_package
 from . import stock_warehouse
+from . import stock_warehouse_orderpoint
