+++ addons/udes_stock/models/stock_location_route.py
@@ -1,0 +1,8 @@
+from odoo import fields, models
+
+
+class Route(models.Model):
+    _inherit = "stock.location.route"
+
+    # Disable translation instead of renaming.
+    name = fields.Char(translate=False)
